requirements.txt
streamlit
yfinance
pandas
numpy
google-generativeai
scikit-learn
plotly
textblob